backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
daphne>=4.0
Pillow>=10.0
